implement the equal sampling interface
